TINY-7829: moved command modules to api
