+++ modules/tinymce/src/core/main/ts/commands/Font.ts
